move PluginIdentity and PluginHandshake out of the internal package
