+++ pkg/plugins/diff/service.go
@@ -99,6 +99,6 @@ func (s *Service) Close() {
 	}
 }
 
-func (s *Service) RegisterDiffClient(diffType string, props internal.HCPluginProperties) {
+func (s *Service) registerDiffClient(diffType string, props internal.HCPluginProperties) {
 	s.pluginHandler.RegisterPlugin(diffType, props)
 }
