+++ pkg/plugins/internal/manager.go
@@ -5,6 +5,8 @@ import (
 	"os"
 	"os/exec"
 
+	"github.com/treeverse/lakefs/pkg/plugins"
+
 	"github.com/hashicorp/go-hclog"
 	"github.com/hashicorp/go-plugin"
 	"github.com/treeverse/lakefs/pkg/logging"
@@ -17,23 +19,6 @@ var (
 
 var allowedProtocols = []plugin.Protocol{
 	plugin.ProtocolGRPC,
-}
-
-// PluginIdentity identifies the plugin's version and executable location.
-type PluginIdentity struct {
-	ProtocolVersion uint
-	// ExecutableLocation is the full path to the plugin server's executable location
-	ExecutableLocation string
-	// ExecutableArgs is the argument list for the provided plugin executable - optional
-	ExecutableArgs []string
-	// ExecutableEnvVars is the environment variable list for the provided plugin executable - optional
-	ExecutableEnvVars []string
-}
-
-// PluginHandshake includes handshake properties for the plugin.
-type PluginHandshake struct {
-	Key   string
-	Value string
 }
 
 // Handler is an interface used to handle the different plugin implementation.
@@ -64,8 +49,8 @@ func NewManager[T any]() *Manager[T] {
 }
 
 type HCPluginProperties struct {
-	ID        PluginIdentity
-	Handshake PluginHandshake
+	ID        plugins.PluginIdentity
+	Handshake plugins.PluginHandshake
 	P         plugin.Plugin
 }
